% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{alphaSif}
\alias{alphaSif}
\title{Create the alpha_{Si}^{q1...qell}}
\usage{
alphaSif(lambda, sigma)
}
\arguments{
\item{lamda}{an object with same structure than [gamma0f(XX)]}

\item{delta}{an object with same structure than}
}
\description{
Create the alpha_{Si}^{q1...qell}
}
\examples{
set.seed(1)
XX<-Gen_design_variables(N=1000,Q=3,p=4)
gamma0=gamma0f(XX)
delta=Gen_hyper_parameters(XX)$delta
lambda=lambda.j.q1...qell.f(XX,gamma0,delta)
sigma=Gen_hyper_parameters(XX)$sigma
alphaSif(lambda,sigma)
}
