% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{gamma0f}
\alias{gamma0f}
\title{Generate gamma0s}
\usage{
gamma0f(XX)
}
\arguments{
\item{XX}{output from Gendesignvariables}
}
\value{
generated gamma0
}
\description{
Generate gamma0s
}
\examples{
XX<-Gen_design_variables(N=1000,Q=3,p=4)
gamma0f(XX)
}
