% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generateotherpop.R
\name{Gen_hyper_otherparameters}
\alias{Gen_hyper_otherparameters}
\title{Create model hyper-parameters and parameters procedure}
\usage{
Gen_hyper_otherparameters(XX)
}
\arguments{
\item{XX}{an output from [Gen_design_variables]}
}
\description{
Create model hyper-parameters and parameters procedure
}
\examples{
N=1000;Q=2;p=3
XX<-Gen_design_variables(N,Q,p)
Gen_hyper_parameters(XX)
}
