% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{Gen_hyper_parameters_model1}
\alias{Gen_hyper_parameters_model1}
\title{Create model 1 hyper-parameters and parameters procedure}
\usage{
Gen_hyper_parameters_model1(XX)
}
\arguments{
\item{XX}{an output from [Gen_design_variables]}
}
\description{
Create model 1 hyper-parameters and parameters procedure
}
\examples{
N=1000;Q=2;p=3
XX<-Gen_design_variables(N,Q,p)
Gen_hyper_parameters_model1(XX)
}
