% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{Generate_all}
\alias{Generate_all}
\title{Create model hyper-parameters and parameters procedure}
\usage{
Generate_all(
  N = NULL,
  Q = NULL,
  p = NULL,
  K_q = sample(2:p, Q, replace = T),
  XX = Gen_design_variables(N, Q, p, K_q),
  hyper = Gen_hyper_parameters(XX),
  nrep = 3
)
}
\arguments{
\item{XX}{an output from [Gen_design_variables]}
}
\description{
Create model hyper-parameters and parameters procedure
}
\examples{
N=1000;Q=3;p=3
GG<-Generate_all(N,Q,p)
}
