% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{lambda.j.q1...qell.f}
\alias{lambda.j.q1...qell.f}
\title{Create the lambda_{j}^{q1...qell} 
lambda_{k1...kell}^{q1...qell}  can be deduced from the lambda_{S_i}^{q1...qell}
If for example, Q=3, ell=2, q_1=1, q_2=2, k_1=k_2=1, then there will be many strata S_i
that correspond to X_{q_1}=1, X_{q_2}=1, and all of them will have the same value for 
lambda_{S_i}^{q1=1...q2=2}}
\usage{
lambda.j.q1...qell.f(XX, gamma0, delta)
}
\arguments{
\item{XX}{an output from [Gen_design_variables]}

\item{gamma0}{an object with same structure than [gamma0f(XX)]}

\item{delta}{an object with same structure than [Gen_hyper_parameters(XX)$delta]}
}
\description{
Create the lambda_{j}^{q1...qell} 
lambda_{k1...kell}^{q1...qell}  can be deduced from the lambda_{S_i}^{q1...qell}
If for example, Q=3, ell=2, q_1=1, q_2=2, k_1=k_2=1, then there will be many strata S_i
that correspond to X_{q_1}=1, X_{q_2}=1, and all of them will have the same value for 
lambda_{S_i}^{q1=1...q2=2}
}
\examples{
XX<-Gen_design_variables(N=1000,Q=3,p=4)
gamma0=gamma0f(XX)
delta=Gen_hyper_parameters(XX)$delta
lambda.j.q1...qell.f(XX,gamma0,delta)
}
