% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generatepop.R
\name{lambdaf}
\alias{lambdaf}
\title{Create the lambda_{k1...kell}}
\usage{
lambdaf(XX, gamma0, delta)
}
\arguments{
\item{XX}{an output from [Gen_design_variables]}

\item{gamma0}{an object with same structure than [gamma0f(XX)]}

\item{delta}{an object with same structure than [Gen_hyper_parameters(XX)$delta]}
}
\description{
Create the lambda_{k1...kell}
}
\examples{
XX<-Gen_design_variables(N=1000,Q=3,p=4)
gamma0=gamma0f(XX)
delta=Gen_hyper_parameters(XX)$delta
lambda.j.q1...qell.f(XX,gamma0,delta)
#' lambdaf(XX,gamma0,delta)
}
