% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesiancomputations.R
\name{model.text}
\alias{model.text}
\title{Create model file}
\usage{
model.text(GG)
}
\arguments{
\item{GG}{an object generated by the function Generate_all}
}
\description{
Create model file
}
\details{
Generates X, Stratum indicator, computes N_j's and model matrix.
}
\examples{
cat(model.text(Generate_all(N=1000,Q=2,p=5)))
}
