% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesiancomputations.R
\name{Trangucci.fit}
\alias{Trangucci.fit}
\title{Compute posterior of all parameters}
\usage{
Trangucci.fit(GG, initf = "random")
}
\arguments{
\item{GG}{an object generated by the function Generate_all}

\item{initf}{a character string, either "arbitrary" or "real values"}
}
\description{
Compute posterior of all parameters
}
\details{
Generates X, Stratum indicator, computes N_j's and model matrix.
}
\examples{
Trangucci.fit(Generate_all(N=1000,Q=2,p=5))
}
